+++ NEWS.md
@@ -6,6 +6,15 @@ width: 128px;
 border-radius: 128px;
 " />
 
+## v1.12.1
+
+- Features
+    - Fail2Ban support with a dedicated log. See README.
+- Fixes
+    - Comic filename parsing improvements.
+    - Downgrade noisy profile forbidden WARNINGS to DEBUG. They are a normal
+      part of connecting.
+
 ## v1.12.0
 
 - Features